chore: move helper to shared utils
